+++ lib/Redesign/RecipePages/Recipe/second_row_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'ComponentsRedesign.dart';
+import '../../components/ComponentsRedesign.dart';
 
 class SecondRowWidget extends StatelessWidget {
   const SecondRowWidget({
